Move all channel-related functions to lib_channels
